+++ packages/atomic/src/components/atomic-tab/atomic-tab.tsx
@@ -10,7 +10,7 @@ import {
  */
 @Component({
   tag: 'atomic-tab',
-  styleUrl: 'atomic-tab.css',
+  styleUrl: 'atomic-tab.pcss',
   shadow: true,
 })
 export class AtomicTab {
@@ -59,7 +59,7 @@ export class AtomicTab {
       <span {...activePart}>
         <button
           part="tab-button"
-          class={'box-border p-2 bg-primary border-none h-10 ' + activeClass}
+          class={'p-2 bg-primary border-none h-10 ' + activeClass}
           onClick={() => this.handleClick()}
         >
           <slot />