+++ packages/atomic/tailwind.config.js
@@ -1,15 +1,20 @@
 module.exports = {
   purge: {
     content: [
-      './src/**/*.tsx',
-      './src/**/*.css'
+      './src/**/*.tsx', './src/index.html', './src/**/*.css'
     ]
   },
   darkMode: false, // or 'media' or 'class'
   theme: {
     extend: {
       colors: {
-        primary: "var(--color-primary)"
+        "primary": "var(--background)",
+        "on-primary": "var(--on-primary)",
+        "secondary": "var(--on-secondary)",
+        "background": "var(--background)",
+        "on-background": "var(--on-background)",
+        "error": "var(--error)",
+        "on-error": "var(--on-error)",
       }
     },
     backgroundColor: (theme) => ({